# feat(benchmark): Complete Issue #109 Phase 3 - Large Data Performance Optimization